+++ CHANGELOG.md
@@ -11,11 +11,13 @@
 * Add file edit links
 * Add filing fee service charge
 * Add filing fees to checkout
+* Add mailing fees
 * Add order summary
 * Add order_date, last_update details
 * Add page count and delivery ids
 * Add point of contact details to delivery record
 * Add pretty_print_delivery_details
+* Add rural delivery fees
 * Apply document fees
 * Disable donate and premium links
 * Disable page count bits